backend/requirements.txt
torch
torchvision
pyyaml
pillow
scikit-learn
numpy
opencv-python
ultralytics
tensorflow
easyocr
fastapi
uvicorn
pix2tex